修复获取部分群失败 修改超时时间 guzzle verify 默认false 修复cookie文件不存在bug

+++ src/Core/Server.php
@@ -73,7 +73,9 @@ class Server
     private function cleanCookies()
     {
         $this->vbot->console->log('cleaning useless cookies.');
-        unlink($this->vbot->config['cookie_file']);
+        if(is_file($this->vbot->config['cookie_file'])){
+            unlink($this->vbot->config['cookie_file']);
+        }
     }
 
     /**
